+++ libgeda/include/libgeda/prototype.h
@@ -25,7 +25,7 @@ BEGIN_DECLS
          char   *f_get_bitmap_filespec           (const char *filename);
        GSList   *f_get_dir_list_files            (      char *path, char *filter);
          char   *f_get_dirname                   (const char *filename);
-         bool    f_get_file_contents             (const char *filename, char **contents, unsigned int *length, GError **err);
+         bool    f_get_file_contents             (const char *filename, char **contents, size_t *length, GError **err);
    const char   *f_get_filename_ext              (const char *filename);
    const char   *f_get_format_header             (void);
          bool    f_get_is_path_absolute          (const char *filename);
